+++ src/utils.rs
@@ -1,10 +1,10 @@
-use quote::Tokens;
+use proc_macro2::{Span, TokenStream};
 use syn::{parse_str, Field, FieldsNamed, FieldsUnnamed, GenericParam, Generics, Ident, Index,
           Type, TypeParamBound, WhereClause};
 
 pub fn numbered_vars(count: usize, prefix: &str) -> Vec<Ident> {
     (0..count)
-        .map(|i| Ident::from(format!("__{}{}", prefix, i)))
+        .map(|i| Ident::new(&format!("__{}{}", prefix, i), Span::call_site()))
         .collect()
 }
 
@@ -50,7 +50,7 @@ pub fn add_extra_ty_param_bound_op<'a>(generics: &'a Generics, trait_ident: &'a 
     add_extra_ty_param_bound(generics, &quote!(::std::ops::#trait_ident))
 }
 
-pub fn add_extra_ty_param_bound<'a>(generics: &'a Generics, bound: &'a Tokens) -> Generics {
+pub fn add_extra_ty_param_bound<'a>(generics: &'a Generics, bound: &'a TokenStream) -> Generics {
     let mut generics = generics.clone();
     let bound: TypeParamBound = parse_str(&bound.to_string()).unwrap();
     for type_param in &mut generics.type_params_mut() {
@@ -60,7 +60,7 @@ pub fn add_extra_ty_param_bound<'a>(generics: &'a Generics, bound: &'a Tokens) -
     generics
 }
 
-pub fn add_extra_generic_param(generics: &Generics, generic_param: Tokens) -> Generics {
+pub fn add_extra_generic_param(generics: &Generics, generic_param: TokenStream) -> Generics {
     let generic_param: GenericParam = parse_str(&generic_param.to_string()).unwrap();
     let mut generics = generics.clone();
     generics.params.push(generic_param);
@@ -68,7 +68,7 @@ pub fn add_extra_generic_param(generics: &Generics, generic_param: Tokens) -> Ge
     generics
 }
 
-pub fn add_extra_where_clauses(generics: &Generics, type_where_clauses: Tokens) -> Generics {
+pub fn add_extra_where_clauses(generics: &Generics, type_where_clauses: TokenStream) -> Generics {
     let mut type_where_clauses: WhereClause = parse_str(&type_where_clauses.to_string()).unwrap();
     let mut new_generics = generics.clone();
     if let Some(old_where) = new_generics.where_clause {
@@ -83,7 +83,7 @@ pub fn add_where_clauses_for_new_ident<'a>(
     generics: &'a Generics,
     fields: &[&'a Field],
     type_ident: &Ident,
-    type_where_clauses: Tokens,
+    type_where_clauses: TokenStream,
 ) -> Generics {
     let generic_param = if fields.len() > 1 {
         quote!(#type_ident: ::std::marker::Copy)